Move SpecificPrice command and handlers into a product sub domain
