+++ src/Adapter/Product/SpecificPrice/Update/SpecificPricePriorityUpdater.php
@@ -26,12 +26,12 @@
 
 declare(strict_types=1);
 
-namespace PrestaShop\PrestaShop\Adapter\Product\Update;
+namespace PrestaShop\PrestaShop\Adapter\Product\SpecificPrice\Update;
 
 use PrestaShop\PrestaShop\Adapter\AbstractObjectModelRepository;
+use PrestaShop\PrestaShop\Core\Domain\Product\SpecificPrice\ValueObject\PriorityList;
 use PrestaShop\PrestaShop\Core\Domain\Product\ValueObject\ProductId;
 use PrestaShop\PrestaShop\Core\Domain\SpecificPrice\Exception\CannotSetSpecificPricePrioritiesException;
-use PrestaShop\PrestaShop\Core\Domain\SpecificPrice\ValueObject\PriorityList;
 use PrestaShop\PrestaShop\Core\Exception\CoreException;
 use PrestaShopException;
 use SpecificPrice;
