+++ src/Core/Domain/Product/SpecificPrice/ValueObject/PriorityList.php
@@ -26,7 +26,7 @@
 
 declare(strict_types=1);
 
-namespace PrestaShop\PrestaShop\Core\Domain\SpecificPrice\ValueObject;
+namespace PrestaShop\PrestaShop\Core\Domain\Product\SpecificPrice\ValueObject;
 
 use PrestaShop\PrestaShop\Core\Domain\SpecificPrice\Exception\SpecificPriceConstraintException;
