+++ tests/Integration/Behaviour/Features/Scenario/Product/SpecificPrice/update_specific_prices.feature
@@ -1,6 +1,6 @@
-# ./vendor/bin/behat -c tests/Integration/Behaviour/behat.yml -s product --tags product-specific-prices
+# ./vendor/bin/behat -c tests/Integration/Behaviour/behat.yml -s product --tags update-specific-prices
 @reset-database-before-feature
-@product-specific-prices
+@update-specific-prices
 Feature: Update product options from Back Office (BO)
   As a BO user
   I need to be able to update product options from BO
